+++ packages/fetch/src/index.ts
@@ -1,5 +1,5 @@
-export * from './Error/FetchError';
-export * from './Mock/MockFetch';
+export * from './a/MockFetch';
 export * from './Fetch';
+export * from './FetchError';
 export * from './IFetch';
 export * from './FetchResponse';
