+++ packages/heap/src/index.ts
@@ -1,4 +1,4 @@
-export * from './Error/HeapError';
-export * from './Mock/MockHeap';
-export * from './IHeap';
+export * from './a/MockHeap';
 export * from './Heap';
+export * from './HeapError';
+export * from './IHeap';
