+++ packages/mysql/src/MySQL.ts
@@ -1,9 +1,9 @@
 import { Inconnu, Kind, Nullable, ObjectLiteral, Reject, Resolve } from '@jamashita/anden-type';
 import { Connection as Conn, createPool, MysqlError, Pool, PoolConfig, PoolConnection } from 'mysql';
 import { Connection } from './Connection';
-import { MySQLError } from './Error/MySQLError';
 import { IMySQL } from './IMySQL';
 import { ITransaction } from './ITransaction';
+import { MySQLError } from './MySQLError';
 
 export class MySQL implements IMySQL {
   private readonly pool: Pool;
