+++ packages/mysql/src/index.ts
@@ -1,8 +1,8 @@
-export * from './Error/MySQLError';
-export * from './Mock/index';
+export * from './a/index';
 export * from './Connection';
 export * from './IConnection';
 export * from './IMySQL';
 export * from './ISQL';
 export * from './ITransaction';
 export * from './MySQL';
+export * from './MySQLError';
