+++ packages/redis/src/index.ts
@@ -1,11 +1,11 @@
-export * from './Error/RedisError';
-export * from './Mock/index';
+export * from './a/index';
 export * from './IRedis';
 export * from './IRedisHash';
 export * from './IRedisList';
 export * from './IRedisSet';
 export * from './IRedisString';
 export * from './Redis';
+export * from './RedisError';
 export * from './RedisHash';
 export * from './RedisList';
 export * from './RedisSet';
