+++ packages/request/src/index.ts
@@ -1,5 +1,5 @@
-export * from './Error/RequestError';
-export * from './Mock/MockRequest';
+export * from './a/MockRequest';
 export * from './IRequest';
 export * from './Request';
+export * from './RequestError';
 export * from './RequestResponse';
